requirements.txt
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
numpyro[cuda]==0.11.0
funsor==0.4.5
cloudpickle==2.2.1
sigfig==1.3.2
scipy==1.10.0
numpy==1.24.1
opt-einsum==3.3.0
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
pyper==1.1.2
tqdm==4.64.1
matplotlib==3.6.3
arviz==0.14.0
seaborn==0.12.2
ipywidgets==8.0.4
widgetsnbextension==4.0.5
pandas-profiling=3.2.0
jupyterlab